requirements.txt
boto==2.21.2
Django==1.6.1
django-athumb==2.0
django-bitly==0.7
django-endless-pagination==2.0
django-imagekit==3.0.4
django-social-auth==0.7.28
django-storages==1.1.8
django-suit==0.2.5
django-suit-redactor==0.0.2
django-taggit==0.11.2
django-taggit-autosuggest==0.2.1
MySQL-python==1.2.3
pillow==2.3.0
python-oauth2==0.4.0
python-openid==2.2.5
pytz==2013.9
South==0.8.2